pyproject.toml
[project]
name = "nonebot_plugin_chatrecorder"
version = "0.7.0"
description = "适用于 Nonebot2 的聊天记录插件"
authors = [{ name = "meetwq", email = "meetwq@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "nonebot2>=2.4.0,<3.0.0",
  "nonebot-plugin-orm>=0.7.7,<1.0.0",
  "nonebot-plugin-uninfo>=0.9.0,<1.0.0",
  "nonebot-plugin-localstore>=0.6.0,<1.0.0",
]

[project.urls]
Homepage = "https://github.com/noneplugin/nonebot-plugin-chatrecorder"
Repository = "https://github.com/noneplugin/nonebot-plugin-chatrecorder"

[dependency-groups]
dev = [
  "nonebot2[fastapi]>=2.4.0,<3.0.0",
  "nonebot-plugin-orm[default]>=0.7.7,<1.0.0",
  "nonebot-plugin-datastore>=1.0.0,<2.0.0",
  "nonebot-session-to-uninfo>=0.0.2,<0.1.0",
  { include-group = "test" },
  { include-group = "adapters" },
]
test = [
  "nonebug>=0.4.3,<0.5.0",
  "pytest>=8.0.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "pytest-xdist>=3.0.2,<4.0.0",
  "pytest-asyncio>=1.0.0,<2.0.0",
  "asyncpg>=0.31.0,<0.32.0",
  "aiomysql>=0.2.0,<0.3.0",
]
adapters = [
  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
  "nonebot-adapter-console>=0.9.0,<0.10.0",
  "nonebot-adapter-kaiheila>=0.3.4,<0.4.0",
  "nonebot-adapter-telegram>=0.1.0b20,<0.2.0",
  "nonebot-adapter-feishu>=2.6.2,<3.0.0",
  "nonebot-adapter-discord>=0.1.8,<0.2.0",
  "nonebot-adapter-dodo>=0.2.1,<0.3.0",
  "nonebot-adapter-satori>=0.13.4,<0.14.0",
  "nonebot-adapter-qq>=1.6.2,<2.0.0",
  "nonebot-adapter-milky>=1.1.0,<2.0.0",
]

[tool.uv]
resolution = "lowest-direct"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = ["ignore::DeprecationWarning", "ignore::RuntimeWarning"]

[tool.nonebot]
plugins = ["nonebot_plugin_chatrecorder"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "Q"]
ignore = ["E402", "E501", "E711", "C901", "UP037"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@overload",
  "except ImportError:",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
